+++ front/src/app/modules/admin-docker-image/interfaces/shutdown-status.ts
@@ -1,0 +1,4 @@
+export interface IShutdownStatus {
+  error: string | null;
+  isShutdown: boolean;
+}