+++ front/CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+## 2.28.2
+### Fixed
+- Improve test environment loader in build image section to stop only when container is available.
+
+### Added
+- Add Shutdown button for test environment
+
 ## 2.28.1
 ### Added
 - Add a information and contact button on the login page